Rework upgrade, remove the 'if upgrade_type == UPGRADE_APP' test.
Indeed, even if the upgrade is only a 'ynh revision', this revision might actually be fixing
what *is* inside this if test.
So we can't discriminate generic things in the script according to the upgrade_type.

The ynh_check_app_version_changed helper is now a simple shim that reads YNH_APP_UPGRADE_TYPE.
The rest of its old logic is now handled by the core.
So instead of calling it, we can just read YNH_APP_UPGRADE_TYPE and add some documentation about
possible values.